requirements.txt
dash==2.12.1
dash.ly==0.17.3
dash_auth==2.0.0
dash_bootstrap_components==1.2.1
gspread==5.7.2
oauth2client==4.1.3
pandas==1.5.2
plotly==5.9.0
PyMySQL==1.0.3
python_dateutil==2.8.2
requests==2.28.1
gunicorn
dash-tools
